+++ simulators/lean/clients/devnet4.yaml
@@ -12,6 +12,5 @@
   nametag: devnet4
 - client: zeam
   nametag: devnet4
-  dockerfile: devnet4
 - client: qlean
   nametag: devnet4
