+++ simulators/lean/helper/lean_spec_client_runner.py
@@ -625,6 +625,20 @@ def encode_enr_port(port: int) -> bytes:
     return port.to_bytes(byte_length, "big")
 
 
+def build_canonical_secp256k1_signature(
+    identity_key: IdentityKeypair,
+    digest: bytes,
+) -> bytes:
+    der_signature = identity_key.private_key.sign(
+        digest,
+        ec.ECDSA(Prehashed(hashes.SHA256())),
+    )
+    signature_r, signature_s = decode_dss_signature(der_signature)
+    if signature_s > SECP256K1_ORDER // 2:
+        signature_s = SECP256K1_ORDER - signature_s
+    return signature_r.to_bytes(32, "big") + signature_s.to_bytes(32, "big")
+
+
 def build_helper_bootnode_enr(identity_key: IdentityKeypair) -> str:
     advertise_ip = os.environ.get(HELPER_ADVERTISE_IP_ENVIRONMENT_VARIABLE)
     if not advertise_ip:
@@ -641,16 +655,8 @@ def build_helper_bootnode_enr(identity_key: IdentityKeypair) -> str:
     }
     unsigned_enr = ENR(signature=b"\x00" * 64, seq=1, pairs=pairs)
     digest = keccak.new(digest_bits=256, data=unsigned_enr._content_rlp()).digest()
-    der_signature = identity_key.private_key.sign(
-        digest,
-        ec.ECDSA(Prehashed(hashes.SHA256())),
-    )
-    signature_r, signature_s = decode_dss_signature(der_signature)
-    # Zeam rejects non-canonical high-s ENR signatures during startup.
     # Normalize helper ECDSA signatures so every client sees a canonical ENR.
-    if signature_s > SECP256K1_ORDER // 2:
-        signature_s = SECP256K1_ORDER - signature_s
-    signature = signature_r.to_bytes(32, "big") + signature_s.to_bytes(32, "big")
+    signature = build_canonical_secp256k1_signature(identity_key, digest)
     return ENR(signature=signature, seq=1, pairs=pairs).to_string()
