+++ simulators/lean/src/scenarios/helper.rs
@@ -2,6 +2,7 @@
 
 use std::collections::HashMap;
 use std::env;
+use std::fmt;
 use std::fs;
 use std::future::Future;
 use std::net::{IpAddr, UdpSocket};
@@ -10,17 +11,18 @@ use std::pin::Pin;
 use std::process::{Child, Command, Stdio};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-use crate::{
-    get_json_with_retry, lean_api_url, lean_environment, selected_lean_devnet, CheckpointResponse,
-    LeanDevnet,
-};
 use alloy_primitives::B256;
 use hivesim::types::ClientDefinition;
-use hivesim::{types::TestResult, Client, Test};
+use hivesim::{types::TestResult, Client, Simulation, Test};
 use reqwest::{header::ACCEPT, Client as HttpClient, Url};
-use serde::Deserialize;
+use serde::{de::DeserializeOwned, Deserialize};
 use tokio::time::{sleep, timeout};
 
+const HIVE_CHECK_LIVE_PORT: &str = "HIVE_CHECK_LIVE_PORT";
+const HIVE_LEAN_DEVNET_LABEL: &str = "HIVE_LEAN_DEVNET_LABEL";
+const LEAN_HTTP_PORT: u16 = 5052;
+const LEAN_DEVNET_CONFIG_PATH: &str = "/app/hive/lean-devnets.txt";
+const LEAN_ROLE: &str = "lean";
 const CHECKPOINT_SYNC_URL_ENVIRONMENT_VARIABLE: &str = "HIVE_CHECKPOINT_SYNC_URL";
 const BOOTNODES_ENVIRONMENT_VARIABLE: &str = "HIVE_BOOTNODES";
 const LEAN_GENESIS_VALIDATORS_ENVIRONMENT_VARIABLE: &str = "HIVE_LEAN_GENESIS_VALIDATORS";
@@ -66,6 +68,61 @@ const LOCAL_HELPER_ENTRYPOINT: &str = "/app/hive/leanspec-client.sh";
 const LOCAL_HELPER_KIND: &str = "lean-spec-local-helper";
 const CLIENT_RUNTIME_ASSET_PREPARER: &str = "/app/hive/prepare_lean_client_assets.py";
 const SSZ_CONTENT_TYPE: &str = "application/octet-stream";
+pub(crate) const HEALTHY_STATUS: &str = "healthy";
+pub(crate) const LEAN_RPC_SERVICE: &str = "lean-rpc-api";
+
+#[allow(dead_code)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub(crate) enum LeanDevnet {
+    Devnet3,
+    Devnet4,
+}
+
+impl LeanDevnet {
+    const DEFAULT: Self = Self::Devnet3;
+
+    fn as_str(self) -> &'static str {
+        match self {
+            Self::Devnet3 => "devnet3",
+            Self::Devnet4 => "devnet4",
+        }
+    }
+}
+
+impl fmt::Display for LeanDevnet {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter.write_str(self.as_str())
+    }
+}
+
+impl TryFrom<&str> for LeanDevnet {
+    type Error = String;
+
+    fn try_from(label: &str) -> Result<Self, Self::Error> {
+        match label.trim() {
+            "devnet3" => Ok(Self::Devnet3),
+            "devnet4" => Ok(Self::Devnet4),
+            other => Err(format!("unsupported lean devnet label {other:?}")),
+        }
+    }
+}
+
+struct LeanDevnetConfig {
+    default_devnet: LeanDevnet,
+    client_support: HashMap<String, Vec<LeanDevnet>>,
+}
+
+#[derive(Debug, Deserialize)]
+pub(crate) struct HealthResponse {
+    pub(crate) status: String,
+    pub(crate) service: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub(crate) struct CheckpointResponse {
+    pub(crate) slot: u64,
+    pub(crate) root: B256,
+}
 
 #[derive(Debug, Deserialize)]
 struct HelperGenesisValidatorEntry {
@@ -119,6 +176,196 @@ impl PostGenesisSyncContext {
             .load_live_fork_choice()
             .await
             .unwrap_or_else(|err| panic!("{err}"))
+    }
+}
+
+pub(crate) fn lean_clients(clients: Vec<ClientDefinition>) -> Vec<ClientDefinition> {
+    clients
+        .into_iter()
+        .filter(|client| client.meta.roles.iter().any(|role| role == LEAN_ROLE))
+        .collect()
+}
+
+pub(crate) fn selected_lean_devnet() -> LeanDevnet {
+    let label =
+        env::var(HIVE_LEAN_DEVNET_LABEL).unwrap_or_else(|_| LeanDevnet::DEFAULT.to_string());
+    LeanDevnet::try_from(label.as_str()).unwrap_or_else(|err| {
+        panic!(
+            "Unsupported Lean devnet selection in environment variable {HIVE_LEAN_DEVNET_LABEL}: {err}"
+        )
+    })
+}
+
+pub(crate) fn set_selected_lean_devnet(devnet: LeanDevnet) {
+    env::set_var(HIVE_LEAN_DEVNET_LABEL, devnet.as_str());
+}
+
+pub(crate) async fn resolve_selected_lean_devnet(simulation: &Simulation) -> LeanDevnet {
+    let config = load_lean_devnet_config();
+    let clients = lean_clients(simulation.client_types().await);
+    if clients.is_empty() {
+        return config.default_devnet;
+    }
+
+    let mut resolved_devnet = None;
+    let mut selected_labels = Vec::new();
+
+    for client in clients {
+        let (base_name, explicit_devnet) = split_client_devnet_name(&client.name);
+        let devnet = explicit_devnet.unwrap_or(config.default_devnet);
+        let supported_devnets = config.client_support.get(base_name).unwrap_or_else(|| {
+            panic!(
+                "Lean client {} is missing from {}",
+                base_name, LEAN_DEVNET_CONFIG_PATH
+            )
+        });
+
+        if !supported_devnets.contains(&devnet) {
+            let supported = supported_devnets
+                .iter()
+                .map(ToString::to_string)
+                .collect::<Vec<_>>()
+                .join(", ");
+            panic!(
+                "Lean client {} does not support {} according to {} (supported: {})",
+                client.name, devnet, LEAN_DEVNET_CONFIG_PATH, supported,
+            );
+        }
+
+        selected_labels.push(format!("{}={devnet}", client.name));
+        match resolved_devnet {
+            Some(previous) if previous != devnet => {
+                panic!(
+                    "Mixed lean devnets selected in one run: {}",
+                    selected_labels.join(", ")
+                );
+            }
+            Some(_) => {}
+            None => resolved_devnet = Some(devnet),
+        }
+    }
+
+    resolved_devnet.unwrap_or(config.default_devnet)
+}
+
+pub(crate) fn lean_environment() -> HashMap<String, String> {
+    let devnet_label = selected_lean_devnet().to_string();
+    HashMap::from([
+        (HIVE_CHECK_LIVE_PORT.to_string(), LEAN_HTTP_PORT.to_string()),
+        (HIVE_LEAN_DEVNET_LABEL.to_string(), devnet_label),
+    ])
+}
+
+pub(crate) fn lean_api_url(client: &Client, path: &str) -> String {
+    format!("http://{}:{}{}", client.ip, LEAN_HTTP_PORT, path)
+}
+
+pub(crate) async fn get_json_with_retry<T: DeserializeOwned>(http: &HttpClient, url: &str) -> T {
+    let mut last_error = String::new();
+
+    for attempt in 1..=30 {
+        match http.get(url).send().await {
+            Ok(response) => {
+                let status = response.status();
+                if !status.is_success() {
+                    last_error = format!("received HTTP {status} from {url}");
+                } else {
+                    return response.json::<T>().await.unwrap_or_else(|err| {
+                        panic!("Unable to decode response from {url}: {err}")
+                    });
+                }
+            }
+            Err(err) => {
+                last_error = err.to_string();
+            }
+        }
+
+        if attempt < 30 {
+            sleep(Duration::from_secs(1)).await;
+        }
+    }
+
+    panic!("Request to {url} did not succeed after retries: {last_error}");
+}
+
+fn load_lean_devnet_config() -> LeanDevnetConfig {
+    let contents = fs::read_to_string(LEAN_DEVNET_CONFIG_PATH).unwrap_or_else(|err| {
+        panic!("Unable to read lean devnet config {LEAN_DEVNET_CONFIG_PATH}: {err}")
+    });
+
+    let mut default_devnet = None;
+    let mut client_support = HashMap::new();
+
+    for (index, raw_line) in contents.lines().enumerate() {
+        let line_number = index + 1;
+        let line = raw_line
+            .split('#')
+            .next()
+            .expect("split always yields at least one segment")
+            .trim();
+        if line.is_empty() {
+            continue;
+        }
+
+        let (key, value) = line.split_once('=').unwrap_or_else(|| {
+            panic!(
+                "Invalid lean devnet config line {} in {}: expected key=value, got {:?}",
+                line_number, LEAN_DEVNET_CONFIG_PATH, raw_line
+            )
+        });
+        let key = key.trim();
+        let value = value.trim();
+
+        if key == "default" {
+            default_devnet = Some(LeanDevnet::try_from(value).unwrap_or_else(|err| {
+                panic!(
+                    "Invalid default lean devnet in {} line {}: {}",
+                    LEAN_DEVNET_CONFIG_PATH, line_number, err
+                )
+            }));
+            continue;
+        }
+
+        let supported_devnets = value
+            .split(',')
+            .map(|label| {
+                LeanDevnet::try_from(label).unwrap_or_else(|err| {
+                    panic!(
+                        "Invalid lean devnet in {} line {} for client {}: {}",
+                        LEAN_DEVNET_CONFIG_PATH, line_number, key, err
+                    )
+                })
+            })
+            .collect::<Vec<_>>();
+
+        if supported_devnets.is_empty() {
+            panic!(
+                "Lean devnet config {} line {} defines client {} without any supported devnets",
+                LEAN_DEVNET_CONFIG_PATH, line_number, key
+            );
+        }
+
+        client_support.insert(key.to_string(), supported_devnets);
+    }
+
+    LeanDevnetConfig {
+        default_devnet: default_devnet.unwrap_or_else(|| {
+            panic!(
+                "Lean devnet config {} must define a default=<devnet> entry",
+                LEAN_DEVNET_CONFIG_PATH
+            )
+        }),
+        client_support,
+    }
+}
+
+fn split_client_devnet_name(client_name: &str) -> (&str, Option<LeanDevnet>) {
+    let Some((base_name, suffix)) = client_name.rsplit_once('_') else {
+        return (client_name, None);
+    };
+    match LeanDevnet::try_from(suffix) {
+        Ok(devnet) => (base_name, Some(devnet)),
+        Err(_) => (client_name, None),
     }
 }
 
@@ -344,9 +591,9 @@ impl RunningLocalLeanSpecHelper {
 
 impl Drop for RunningLocalLeanSpecHelper {
     fn drop(&mut self) {
-        let _ = self.child.kill();
-        let _ = self.child.wait();
-        let _ = fs::remove_dir_all(&self.asset_root);
+        self.child.kill().ok();
+        self.child.wait().ok();
+        fs::remove_dir_all(&self.asset_root).ok();
     }
 }
 
@@ -642,7 +889,7 @@ pub(crate) async fn run_data_test_with_timeout<T: Send + 'static>(
         }
         Err(_) => {
             join_handle.abort();
-            let _ = join_handle.await;
+            join_handle.await.ok();
             TestResult {
                 pass: false,
                 details: format!(
@@ -674,9 +921,8 @@ fn adjusted_genesis_time_for_helper_mesh(
         return requested_genesis_time;
     }
 
-    // Mesh helpers do not backfill the earliest chain history on startup, so they
-    // need a fresh pre-genesis launch window to observe the canonical chain from
-    // the beginning instead of joining after the source helper is already live.
+    // Mesh helpers don't backfill the earliest chain history on startup so they need a fresh pre-genesis launch window to observe the canonical chain from
+    // the beginning instead of joining after the source helper is already live
     let helper_startup_buffer =
         ((helper_peer_count - 1) as u64) * MESH_HELPER_GENESIS_BUFFER_PER_PEER_SECS;
     requested_genesis_time.max(default_genesis_time() + helper_startup_buffer)
@@ -780,13 +1026,12 @@ pub(crate) async fn start_post_genesis_sync_context(
                         test_data.client_under_test.name
                     )
                 });
-                    let _ = test
-                        .start_client_with_files(
-                            test_data.client_under_test.name.clone(),
-                            Some(initial_client_under_test_environment.clone()),
-                            Some(files),
-                        )
-                        .await;
+                    test.start_client_with_files(
+                        test_data.client_under_test.name.clone(),
+                        Some(initial_client_under_test_environment.clone()),
+                        Some(files),
+                    )
+                    .await;
                 }
 
                 panic!("{err}");
@@ -1129,7 +1374,7 @@ pub(crate) fn prepare_client_runtime_files(
     if !output.status.success() {
         let stdout = String::from_utf8_lossy(&output.stdout);
         let stderr = String::from_utf8_lossy(&output.stderr);
-        let _ = fs::remove_dir_all(&local_root);
+        fs::remove_dir_all(&local_root).ok();
         return Err(format!(
             "{CLIENT_RUNTIME_ASSET_PREPARER} failed for {client_type} with status {}.\nstdout:\n{}\nstderr:\n{}",
             output.status, stdout, stderr
@@ -1138,7 +1383,7 @@ pub(crate) fn prepare_client_runtime_files(
 
     let mut files = HashMap::new();
     collect_client_runtime_files(&local_root, &local_root, &container_root, &mut files)?;
-    let _ = fs::remove_dir_all(&local_root);
+    fs::remove_dir_all(&local_root).ok();
     Ok(files)
 }
 
@@ -1179,7 +1424,7 @@ async fn start_client_under_test_attempt(
         }
         Err(_) => {
             handle.abort();
-            let _ = handle.await;
+            handle.await.ok();
             Err(format!(
                 "startup attempt exceeded {} seconds",
                 CLIENT_UNDER_TEST_STARTUP_ATTEMPT_TIMEOUT_SECS
@@ -1252,7 +1497,7 @@ async fn start_checkpoint_sync_client_under_test_with_retry(
     let mut last_error = None;
 
     for attempt in 1..=CLIENT_UNDER_TEST_STARTUP_ATTEMPTS {
-        let _ = ensure_checkpoint_sync_source_ready(
+        ensure_checkpoint_sync_source_ready(
             params.helper,
             params.source_fork_choice,
             &params.helper_config,
@@ -1978,7 +2223,16 @@ mod tests {
                 match listener.accept() {
                     Ok((mut stream, _)) => {
                         let mut request_buffer = [0_u8; 1024];
-                        let _ = stream.read(&mut request_buffer);
+                        let bytes_read = match stream.read(&mut request_buffer) {
+                            Ok(bytes_read) => bytes_read,
+                            Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => 0,
+                            Err(error) => {
+                                panic!("test metadata server should read request: {error}")
+                            }
+                        };
+                        if bytes_read == 0 {
+                            thread::yield_now();
+                        }
 
                         let body = format!(
                             "{{\"genesis_time\":{},\"genesis_validator_entries\":[{{\"attestation_public_key\":\"0xabc\",\"proposal_public_key\":null}}]}}",
