+++ simulators/lean/src/scenarios/rpc_compat.rs
@@ -1,14 +1,12 @@
 use crate::scenarios::helper::{
-    default_genesis_time, expect_single_client, http_client, lean_single_client_runtime_setup,
+    default_genesis_time, expect_single_client, get_json_with_retry, http_client, lean_api_url,
+    lean_clients, lean_environment, lean_single_client_runtime_setup,
     lean_single_client_runtime_setup_with_live_helper, load_fork_choice_response,
     load_response_with_retry, prepare_client_runtime_files, run_data_test_with_timeout,
-    start_post_genesis_sync_context, ClientUnderTestRole, ForkChoiceResponse,
-    HelperGossipForkDigestProfile, PostGenesisSyncContext, PostGenesisSyncTestData,
-    TimedDataTestSpec,
-};
-use crate::{
-    get_json_with_retry, lean_api_url, lean_clients, lean_environment, selected_lean_devnet,
-    CheckpointResponse, HealthResponse, LeanDevnet, HEALTHY_STATUS, LEAN_RPC_SERVICE,
+    selected_lean_devnet, start_post_genesis_sync_context, CheckpointResponse, ClientUnderTestRole,
+    ForkChoiceResponse, HealthResponse, HelperGossipForkDigestProfile, LeanDevnet,
+    PostGenesisSyncContext, PostGenesisSyncTestData, TimedDataTestSpec, HEALTHY_STATUS,
+    LEAN_RPC_SERVICE,
 };
 use alloy_primitives::{FixedBytes, B256};
 use hivesim::{
@@ -848,15 +846,11 @@ dyn_async! {
         let client = expect_single_client(clients);
         let http = http_client();
 
-        let checkpoint: CheckpointResponse = get_json_with_retry(
+        get_json_with_retry::<CheckpointResponse>(
             &http,
             &lean_api_url(&client, "/lean/v0/checkpoints/justified"),
         )
         .await;
-
-        let CheckpointResponse { slot, root } = checkpoint;
-        let _ = slot;
-        let _ = root;
     }
 }
